pyproject.toml
[tool.poetry]
name = "fence"
version = "13.0.3"
description = "Gen3 AuthN/AuthZ OIDC Service"
authors = ["CTDS UChicago <cdis@uchicago.edu>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/uc-cdis/fence"
include = [
  "NOTICE",
]

[tool.poetry.dependencies]
python = ">=3.13,<4.0.0"
alembic = ">=1.7.7"
authlib = ">=1.6.5"
authutils = ">=7.2.2"
bcrypt = ">=3.1.4"
boto3 = "*"
botocore = "*"
cached_property = ">=1.5.1"
cdiserrors = "<3.0.0"
cdislogging = ">=1.2.0"
cdispyutils = ">=2.0.1"
defusedxml = "<1"
flask = ">=3.0.0"
cryptography = ">=44.0.1"
flask-cors = "<7"
flask-restful = ">=0.3.8"
email_validator = ">=1.1.1"
gen3authz = ">=3.1.0"
gen3cirrus = ">=4.0.1"
gen3config = "<3"
gen3users = ">=1.0.2"
markdown = ">=3.1.1"
paramiko = ">=2.6.0"
prometheus-client = "<1"
psycopg2-binary = "<3"
gunicorn = ">=22.0.0"
PyJWT = ">=2.4.0"
python_dateutil = ">=2.9.0"
python-jose = ">=3.5.0"
pyyaml = ">=6.0.1"
requests = ">=2.32.4,<3"
retry = ">=0.9.2"
sqlalchemy = ">=1.3.3"
userdatamodel = ">=3,<4"
werkzeug = ">=3.0.0"
cachelib = ">=0.2.0"
azure-storage-blob = ">=12.6.0"
# azure-storage-blob sub-dependency azure-core requires a security bump but azure-storage-blob hasn't included it yet.
# Force it for now, try removing this later.
# TODO: Remove this explicit sub-dep when azure-storage-blob is updated beyond 12.28.0 and
#       has azure-core >= 1.38.0
azure-core = ">=1.38.0"

Flask-WTF = ">=1.0.0"
itsdangerous = '*' # Use Flask-WTF's itsdangerous
# NOTE: for testing with updated libaries as git repos:
# foobar = {git = "https://github.com/uc-cdis/some-repo", rev = "feat/test"}
wtforms = ">=3.2.1"
backoff = ">=2"
pytz = ">=2025.2"
mock = ">=2.0.0"
httpx = ">=0.28.1"
urllib3 = ">=2.6.3, <3"

[tool.poetry.group.dev.dependencies]
addict = ">=2.2.1"
cdisutilstest = {git = "https://github.com/uc-cdis/cdisutils-test", tag = "2.0.0"}
codacy-coverage = ">=1.3.11"
coveralls = ">=2.1.1"
moto = ">=1.1.24"
pytest = ">=6.2.5"
pytest-cov = ">=2.5.1"
pytest-flask = ">=1.3.0"
poetry = ">=2.2.1"
deptry = ">=0.23.1"
responses = ">=0.25.8"

[tool.deptry.per_rule_ignores]
DEP002 = [
  "gunicorn",  # not imported but used in Dockerfile
  "psycopg2-binary",   # not imported but used by sqlalchemy
  "urllib3", # Needed to be pinned to a version >=2.6.0

]

[tool.poetry.scripts]
fence-create = 'bin.fence_create:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
